completed tests

+++ LogixGoDesktop/src/utils/Helper.java
@@ -79,7 +79,7 @@ public class Helper {
 
 	public static boolean isNumeric(String str) {
 		try {
-			Integer.parseInt(str);
+			Long.parseLong(str);
 			return true;
 		} catch (NumberFormatException e) {
 			return false;
